Fix ExtensionContainer retrieval logic
[Problem]
For shared state operations, ExtensionContainer that is to be operated
on is being fetched from the registered extension map directly using the
"extensionName". This is incorrect because  registered extension map maintains a mapping
between the class name of the extension and the container, resulting in a no-op always.

[Solution]
Add a utility method to get the extension container form the registered extension map
based on the extension name.