Fix bundled rules extraction logic
[Problem]
Currently bundled rules are extracted to `adbdownloadcache`
but the logic incorectly assumes that they are present in a sub-directory
structure `adbdownloadcache/ADB-MobileConfigRules` resuting in the bundled
rules application logic to fail.

[Fix]
Fix the directory structure to use the correct location. Use
`adbdownloadcache/configRules/bundledRules` to be the location where the
bundledRules are extracted. This will ensure consistancy with the location
where cached rules are downloaded extracted.