Fix race condition in SerialWorkDispatcher
SerialWorkDispatcher currently ensures that one and only
one WorkProcessor is active and sifting through the queue of
work items. This is done so by checking if the future associated
with the work processor is done.

However, there is a case where the WorkProcessor is about to be done
and completion check is done on the future referring it, resulting in
the dispatcher being stuck until a new work item is added to resume the
processor again.

Fix the race by eagerly resuming right before the WorkProcessor exits
when there are items to process.