Add optional serializing of event data
Currently when event data is set and unsuported types
are encountered the entire event data is set to be empty.

There are some cases where we automatically collect data
from activity launches and it is not always the case that the
intent has supported types. One such case is being seen with
deeplinks collected during target previews where the event data
is being relied upon to show a preview.

Add support for components setting the data to allow skipping
unspported types instead of dropping the entire data map from
an event.