Do not marshal intent extras containing unsupported types
Currently, collectLaunchInfo() automatically attmpts to send
data with the intent uri that launched any activity and its extras
as event data. If the intent extra data contains unsupported types,
 an event with empty event data is sent.

This poses a problem for Target previews where the intent uri is still
needed for showing a preview.

Given that DataMarshaller is only used internally by MobileCore and
it is only used to marshal intent data for sending them as events,
change the behavior of marshaling to the follwing :
- Always add the intent uri when marshaling
- Add intent extras to the event data if it is entirely of supported Event types
- Drop intent extras from Event data if it contains atlest on unsupported type.

This allows limiting the behaviour changes to collectLaunchInfo() and
no data is additionally dropped which would otherwise be sent now.