Align message display notification behavior across API levels
Currently status of whether or not messages are being displayed
on the screen are maintained by MessageMonitor via a notification
mechanism. Control that a message is displayed is tied to onAttach() system
callback.
On API 22 and below the onAttach() callback is being seen only once when
the message is attached to the host activity. So,on an orientation change,
the activity is recreated and the fragment is removed from the hierarchy but
the re-attachment notification is not received. This is causing MessageMonitor
state to be stale (that no message is being displayed) resulting in any operation
on the screen/app to be swallowed.

Changes:
- onActivityCreated() callback is always invoked when fragment is attached to
  activity even on orientation changes. Use this as a mechanism o notify
  MessageMonitor.
- Since the WebView is reused across orientation changes, ensure that it is
  removed from previous parents.
- Shift acivity checking in AEPMessage slightly earlier to avoid false notification
  of messages being displayed.