Fix message viewed reporting
Notification about a message being viewed is never being sent to edge on
API 21/22 and sent multiple times on orientation change on newer devices.
So this cannot be relied upon.

Notify about a message viewed only once immediately after successfully
showing the message in AEPMessage instead of tying it to the figment lifecycle.
This ensures that viewed is only called once.Additionally,
multiple shows() without dismiss is guarded by the MessagingMonitor.