Fix race during js hanlder registration
Javascript handlers should be registered on the main thread and will take effect
on the next load of the webview. InAppMessagePresentable invokes
`DefaultInAppMessageEventHandler#onNewWebView()`and then later loads the url,
both on the main thread. `DefaultInAppMessageEventHandler#onNewWebView()`
posts the javascript handler registration again on the main thread via a coroutine.
This results in the non-zero probability of loading the webview before the handler
registration coroutine is executed on the main thread leading to the native callbacks
failing.

Since the notification of new webview happens on the main thread do no re-post it again
to allow execution of js handler registration and url loading serially.