Force WebView bounds to match AndroidView dimensions
Currently the WebView that renders the InAppMessage is wrapped
inside an AndroidView composable. Leaving it as is is causing
HTML's with "overflow" based CSS manipulation to render incorrectly.

Explicitly make the WebView to match the dimensions of the AndroidView
that wraps it. This allows fixing the bounds of the webview and overflow.