Enable consumers to use default InAppMessage settings
Currently, 2.x In App message users have the ability to
use the the internal webveiw constructed by the SDK and
place them in their view directly instead of the SDK doing it.

We do not want to expose the internal WebView instance to the cusromer
apps in 3.x. Instead expose a way for customers to apply the default
settings to the WebView they create,