Fix crash during composable attachment to android.app.Activity
android.app.Activity, which is the base class in the framework package,
does not provide a lifecycle owner by default.

ComposeView requires a lifecycle owner  when attaching to a view and
not finding one will result in a crash. android.app.Activty has not been
updated for a long time and Android created AppCompatActivity/Fragment activity
in the support package to allow updates to be compatible with old android
devices. Android recommends using inheriting from AppCompatActivity or
similar which provides a lifecycle owner by default (irrespective of whether
compose is being used or not).

There may be cases where apps use this legacy Activity in their apps and launching
our Ui Service components on top of them will not work unless the ComposeView sees that
there is some lifecycle owner.
To fix this, attach a proxy lifecycle owner and related owners
to the view of the activity just before attaching the ComposeView and detach it when
the activity is destroyed. This will allow displaying out UI components like InAppMessages,
Floating buttons on these legacy implementations.

Since there is originally no observer attachable to the Activity, there are no components
dependent on the events from this owner except for the compose view.