Use InAppMessage dimensions relative to host activity
Currently, the size of the InAppMessage is based on the
screen height and width. While this displays the IAM ass desired
for all single activity cases, this is not always the case when
the activity is not occupying the entire screen.
Such cases will see a full screen message within the activity
irrespective of the dimensions of the actual content size of
the activity. So 70% height would be 70% screen height and exceed
the activity content view height where as it should have been
70% of the height within the activity.

Use the Use InAppMessage dimensions relative to host activity
and not the entire screen.