Always recompute screen size related calculations
Currently, screen shize related calculations are done once
and remembered on recompositions. This works well when adapting
to the new size when orientation of the device is changed.

However, when the activity hosting the composable restricts configuration
changes, the remembered value is not recomputed preventing the presentables
from adapting to the new screen size.

As a solution, do not remember the size related calculation. Always calculate
them on composition.