Set webview container background to be transparent
The InAppMessage WebView is hosted inside a Card container.
By default the Card background color is grey. This is causing
unintended results(grey bg being seen) for  in-app messages
where dimensions are set to fill the entire screen but the content
only occupies a portion of it.

As a soution, set the container Card's background color to Transperent
as with the WebView.