Detach presentables before re-attaching to top Activity
Currently, the ComposeView attached to the Activity when
presentables are shown is not removed until the activity
is destroyed.
This is OK for most use-cases that follow the sigle window mode
where the newly resumed activity obscures the previous one, and
the presentable redraws on the new activity. However, this approach
of waiting until the activity is destroyed may result in the
presentables being shown on each activity in a multi-window side-by-side
mode.
While we do not support multi-window mode, we should ensure that the
experience does not deteriorate in such a case. As a solution,
change the current detachment logic to to the following:
 - Track the current activity that the presentable is attached to
 - When presentable is visble on Activity A and another activity B
   resumes, use the Android activity co-ordination guarantees to
   detach presentable from Activity-A and re-attach to Activity B
 - Reset the reference held to the activity that the presentable is
   attached to when the activity is destroyed.