Add support for debug events
The SDKs often needs to change its standard business logic to
enable debugging/testing workflows. We need an event structure that
external systems can follow to generate an event that communicates
to the listening SDK extension that it should operate in a debug
mode - a mode outside of expected production behavior if one exists.

This commit adds an eventSource "com.adobe.eventSource.debug" to
identify such events.
Further, it adds utility methods to identify the eventType and eventSource
provided within the `data.debug` content of such events.