Apply .99 alpha to versions <=22 only
Currently a hack to apply an alpha of .99 on the
content view holding the in-app message to ensure
that corner radius settings are honored for API <=22.

However this is applied to all api levels currently.
A very keen look show a faint outline the contents underneath
the in-app message. Restrict the hack to API <=22 only to
ensure that the issue is not seen on devices where the hack is not needed.

Note that the slight transperency due to .99 alpha will still exist
in API <=22 to honor rounded corner settings.