Prevent CardView from applying a shadow
CardView has a default elevation of 1.dp and  applies a slight shadow
on the background.
While this is not evident when the HTML completely covers the dimension
of the in-app message, it can be seen when HTML only occupies a portion
of the in app message dimension.

Set the elevation to 0.dp for the CardView to prevent any shadows.